Refactorizacion #2
Code Smell: Switch Statement
Tecnicas aplicadas:
-Extract Class
-Replace TypeCode with State/Strategy
-repalce conditional with Polymorfism
