+++ src/main/java/Gestion/GestionCompleta.java
@@ -1,0 +1,16 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
+ */
+package Gestion;
+
+/**
+ *
+ * @author vv
+ */
+public interface GestionCompleta extends GestionParcial{
+    
+    public void eliminar();
+    
+    public void actualizarDatos();
+}
