+++ src/main/java/modelos/Producto.java
@@ -24,6 +24,10 @@ public class Producto {
         this.idProveedor = idProveedor;
     }
 
+    public Producto(int id) {
+        this.id = id;
+    }
+
     public int getId() {
         return id;
     }
@@ -79,4 +83,28 @@ public class Producto {
     public void setIdProveedor(int idProveedor) {
         this.idProveedor = idProveedor;
     }
+
+    @Override
+    public int hashCode() {
+        int hash = 7;
+        hash = 59 * hash + this.id;
+        return hash;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (obj == null) {
+            return false;
+        }
+        if (getClass() != obj.getClass()) {
+            return false;
+        }
+        final Producto other = (Producto) obj;
+        return this.id == other.id;
+    }
+    
+    
 }
