+++ src/main/java/modelos/TipoCuenta.java
@@ -1,0 +1,16 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Enum.java to edit this template
+ */
+package modelos;
+
+/**
+ *
+ * @author vv
+ */
+public enum TipoCuenta {
+    edu,
+    gob,
+    outlook,
+    gmail
+}
